# agent-manager-service/docs/api_v1_openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Agent Manager Service API
servers:
  - url: /api/v1
paths:
...
  /orgs/{orgName}/environments:
    parameters:
      - name: orgName
        in: path
        required: true
        description: Organization name/handle
        schema:
          type: string
          pattern: '^[a-z0-9-]+$'
          minLength: 1
          maxLength: 64

    get:
      tags:
        - Environments
      summary: List all environments
      description: Retrieve a paginated list of all environments for the organization
      operationId: listEnvironments
      parameters:
        - name: limit
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
      responses:
        '200':
          description: Successfully retrieved environment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentListResponse'
        '400':
          description: Bad request - invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Environments
      summary: Create a new environment
      description: |
        Create a new gateway environment for the organization.

        Environment names must be unique within an organization and should be lowercase
        alphanumeric with hyphens (e.g., dev, staging, production).
      operationId: createEnvironment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEnvironmentRequest'
      responses:
        '201':
          description: Environment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayEnvironmentResponse'
        '400':
          description: Bad request - invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - resource already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /orgs/{orgName}/gateways:
    parameters:
      - name: orgName
        in: path
        required: true
        description: Organization name/handle
        schema:
          type: string
          pattern: '^[a-z0-9-]+$'
          minLength: 1
          maxLength: 64

    get:
      tags:
        - Gateways
      summary: List all gateways
      description: |
        Retrieve a paginated list of all gateways for the organization.

        Supports filtering by gateway type, status, and environment.
      operationId: listGateways
      parameters:
        - name: limit
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: type
          in: query
          description: Filter by gateway type
          schema:
            type: string
            enum:
              - AI
              - REGULAR
        - name: status
          in: query
          description: Filter by gateway status
          schema:
            type: string
            enum:
              - ACTIVE
              - INACTIVE
              - PROVISIONING
              - ERROR
        - name: environment
          in: query
          description: Filter by environment name
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved gateway list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayListResponse'
        '400':
          description: Bad request - invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Gateways
      summary: Register a new gateway
      description: |
        Register a new AI gateway with the platform.

        **Gateway Types:**
        - REGULAR: Handles incoming API traffic
        - AI: Handles outgoing LLM/MCP traffic (used for AI agents)

        **Security Note:**
        - Gateway credentials are encrypted before storage
        - Credentials are never returned in API responses
        - Use a strong encryption key configured in GATEWAY_ENCRYPTION_KEY
      operationId: registerGateway
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGatewayRequest'
      responses:
        '201':
          description: Gateway registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayResponse'
        '400':
          description: Bad request - invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - resource already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /orgs/{orgName}/gateways/{gatewayID}/environments:
    parameters:
      - name: orgName
        in: path
        required: true
        description: Organization name/handle
        schema:
          type: string
          pattern: '^[a-z0-9-]+$'
          minLength: 1
          maxLength: 64
      - name: gatewayID
        in: path
        required: true
        description: Gateway UUID or name
        schema:
          type: string

    get:
      tags:
        - Gateway-Environment Mappings
      summary: Get gateway environments
      description: |
        List all environments mapped to a specific gateway.

        This shows which environments can use this gateway for routing traffic.
      operationId: getGatewayEnvironments
      responses:
        '200':
          description: Successfully retrieved environment list
          content:
            application/json:
              schema:
                type: object
                properties:
                  environments:
                    type: array
                    items:
                      $ref: '#/components/schemas/GatewayEnvironmentResponse'
        '400':
          description: Bad request - invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    EnvironmentListResponse:
      type: array
      items:
        $ref: "#/components/schemas/Environment"
      description: List of environments
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Human-readable error message
        description:
          type: string
          description: Error description
        additionalData:
          type: object
          additionalProperties: true
      required:
        - message
...
    CreateEnvironmentRequest:
      type: object
      required:
        - name
        - displayName
        - dataplaneRef
        - dnsPrefix
      properties:
        name:
          type: string
          description: Unique environment name (lowercase, alphanumeric with hyphens)
          pattern: '^[a-z0-9-]+$'
          minLength: 1
          maxLength: 64
          example: production
        displayName:
          type: string
          description: Human-readable display name
          minLength: 1
          maxLength: 128
          example: Production Environment
        description:
          type: string
          description: Optional description of the environment
          example: Production environment for critical AI agents
        dataplaneRef:
          type: string
          description: Reference to the dataplane
          minLength: 1
          maxLength: 100
          example: default
        dnsPrefix:
          type: string
          description: DNS prefix for the environment
          minLength: 1
          maxLength: 100
          example: default
        isProduction:
          type: boolean
          description: Whether this is a production environment
          example: false
...
    GatewayListResponse:
      type: object
      required:
        - gateways
        - total
        - limit
        - offset
      properties:
        gateways:
          type: array
          items:
            $ref: '#/components/schemas/GatewayResponse'
        total:
          type: integer
          format: int32
          description: Total number of gateways matching the filter
          example: 5
        limit:
          type: integer
          format: int32
          description: Number of results per page
          example: 20
        offset:
          type: integer
          format: int32
          description: Number of results skipped
          example: 0
...